+++ Libplanet.Explorer/Store/MySQLRichStore.cs
@@ -140,12 +140,6 @@ namespace Libplanet.Explorer.Store
             return _store.ContainsTransaction(txId);
         }
 
-        /// <inheritdoc cref="IStore.CountTransactions()"/>
-        public long CountTransactions()
-        {
-            return _store.CountTransactions();
-        }
-
         /// <inheritdoc cref="IStore.CountBlocks()"/>
         public long CountBlocks()
         {
@@ -238,23 +232,11 @@ namespace Libplanet.Explorer.Store
         ) =>
             _store.ForkBlockIndexes(sourceChainId, destinationChainId, branchpoint);
 
-        /// <inheritdoc cref="IStore.IterateTransactionIds()"/>
-        public IEnumerable<TxId> IterateTransactionIds()
-        {
-            return _store.IterateTransactionIds();
-        }
-
         /// <inheritdoc cref="IStore.GetTransaction{T}(TxId)"/>
         public Transaction<T> GetTransaction<T>(TxId txid)
             where T : IAction, new()
         {
             return _store.GetTransaction<T>(txid);
-        }
-
-        /// <inheritdoc cref="IStore.DeleteTransaction(TxId)"/>
-        public bool DeleteTransaction(TxId txid)
-        {
-            return _store.DeleteTransaction(txid);
         }
 
         /// <inheritdoc cref="IStore.IterateBlockHashes()"/>